Fix off-by-one error in bfile fibonacci example.
Running examples/chapel.bfile.fib.py produced the output 89 as the 10th
fibonacci number, which is not correct.  That is the 11th fibonacci number.
The error arose from using 0 to n as the traversal (n+1 numbers), instead of 1
to n (n numbers).  While I was there, I removed the unnecessary traversal
variable i, since for loops can function without storing their iteration if it
isn't used by the body.