Fix a few more examples
chapel.sfile.hello.ename.py was also trying to access a file that did not exist,
(chapel.hellolib.exported.chpl), so I updated it to point to the correct file.
c.sfile.py was trying to access a file that did not exist and for which no
equivalent was available, so I created one (sfiles/c/examples.c).  Two tests
were exactly the same (chapel.inline.numpy.example.py and
chapel.inline.numpy.py), so I removed the one with the longer name and added a
comment in the remaining one for my own edification on the expected output.