Use repository hierarchy instead of absolute path for sfile
The test chapel.sfile-inline.py was depending on an absolute path to find the
location of a chapel file that was outside the normal sfile storage location.
The absolute location was both machine and user-specific.  I replaced the path
with a relative path that depends on the directory structure for the repository.
This allows a wider ability to successfully run this test.  It's a little
finicky, but I intend to make the test more robust later.