Update first of the examples that are visible online
Converts chapel.sfile.hellolib.py to test_sfile_hellolib.py and adds test code
to the end of it.  Since we don't really want to display the test code (and I
needed to update the filename anyways), I modified the usage of this example to
only include the relevant line numbers.