Fix output dependent tests
Py.test's capfd and capsys depend on certain circumstances to be true - namely,
that output is placed in a specific place.  With our Jenkins run, that is not
the case - either because of how python is set up to interact with Chapel or how
Chapel is set up, or something.  It doesn't really matter (though I am curious).
It is still useful to have output tests, so instead of trying to depend on
py.test's built in functionality, implement our own method of capture (based in
part on the testing structure we built to ensure chplforpyp-docs examples work)

Because this method captures the initial message about creating a temporary
file for the materialized Chapel code (which only occurs the first time a test
is run), some of the previous assumptions about the output "starting with" a
particular sequence no longer hold.  However, we do want to make sure these
sequences precede the rest of the output, so the tests have been adjusted in
kind.