Two key changes to this branch
First, the some of the numpy examples caused the test following them to segfault
for reasons I have yet to ascertain.  For now, include only the one that does
not have this problem (I reverted the commit which did all of them).  Second,
modify the examples which generate output to reflect what I learned with the
other output tests (namely, that the Jenkins setup is not quite compatible with
how capfd picks up output so a different method must be employed).  Thankfully,
a couple of these tests did not need to be changed.