Incorrect(?) Materialization Error example
These two tests demonstrate an error that pytest had exercised at one point,
which is: if two extern functions are named the same, the first one used will
execute for all attempts to access an extern function by that name (even if the
intended source file of the other differs).  For instance, if I have a function
foo in module1 and a function foo in module2, if module2.foo is different from
module1.foo but module1.foo is called first, when I try to call module2.foo,
the result will be the same as if I had called module1.foo again.

I think this is a bug and that the materialization of functions should take into
account their sources (i.e. inline vs. named bfile vs. named sfile) and if the
source differs from the original materialization, a second materialization
should be made.

If we have support for accessing multiple modules in a Chapel file, it'd be nice
for the separate materialization to apply to functions in separate modules.
That will probably be tricky, though.