Add three tests of using arbitrary Chapel modules within export Chapel code
Unfortunately, right now we cannot include use statements that reference Chapel
code which is not present in the internal or standard modules in our exported
or inlined Chapel functions called from Python.  These tests attempt to do so
by placing the other Chapel file in question in the same sfile/bfile location as
the calling code.

Since Chapel module files such as Time have been used in other tests, I remain
hopeful that we can expand pyChapel to make use of specified locations to look
for further Chapel code - however, it is apparent that work needs to be done to
achieve this (I believe c.library.py demonstrates an in-progress attempt at this
functionality, for instance).