Initial ability to include directories for Chapel to search for modules + tests
This allows the user to specify directories which contain Chapel files upon
which your bfiles and sfile-defined functions depend.  It allows one to use a
module living in the same directory as the sfile you specified, but we would
prefer that to be noticed without needing to specify the directory location.
That is hopefully a near-term next step.  Unfortunately, this fix does not allow
inline function definitions to access outside modules yet.  That is my next
step.

This commit includes the updating of the previous bfile and sfile tests
(test_arbitrary_chapel_use.py and test_bfile_arbitrary_chapel.py), as well as
the addition of two new tests.  One of the new tests exercises using multiple
module directories, and the other exercises the case where a later extern Chapel
function call shares a name with the function utilized within a prior extern
function (which was accessed from a separate module) to make sure that no
naming conflict arises.

Details of the feature:

When declaring a Chapel function from an sfile or a bfile, in the decorator add
a new argument "depend", which takes a list of directories. For example:

@Chapel(depend=[FILEPATHS], sfile=SFILE_NAME)
def foo():

This argument has been added to the interface for Extern and its child class
Chapel (FromC does not take this argument and sends in an empty list) in
pych/extern.py.  The runtime materialization passes the depend field on to the
compiler call for Chapel (I'm going to alter this stage for better modularity,
but it works right now so I wanted to stash that work), which replaces a
__DEPENDS__ notation in the compile line with a '-M MODULENAME' per list entry.
The __DEPENDS__ notation follows similar actions taken with other variables by
that section of the code, we can debate the utility of that course of action at
a different time.