Fix the modularization of my previous approach
Though it proved the fix worked, my prior fix invoked bad style by sending the
details of extern.depend on to the compiler command and making it parse and deal
with the contents.  This commit moves the conversion of the contents into a
form that is usable in the compilation command into the materialization of the
external function, in keeping with the practices visible therein (by which I
mean, knowledge of the inner fields of an Extern class).