Add another test (and some minor cleanup)
This test ensures that we get an error that we can't find the module when we
wouldn't be able to find it without the "depend" argument.

Additionally, I removed references to pytest within the tests that no longer
fail, as they aren't using pytest code now.  I also added a comment on the
assumption made by my "depend" addition.