Update test of failing case.
I foolishly forgot that once a chapel file has been properly processed, it won't
be recompiled if the file hasn't changed.  Since my test of failing to link
tests together comes after the first test of linking together and both were
using the same base chapel file, the "failure" test was executing instead of
generating an error message.  Whoops.  Add a new chapel base file for it to use,
with a note warning people not to make use of that file if they want all tests
to pass, and make the "failure" test point to that file.